requirements.txt
asgiref==3.8.1
astroid==3.2.4
attrs==24.2.0
autobahn==23.1.1
Automat==24.8.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.7.1
dill==0.3.8
Django==4.1
django-channels==0.7.0
django-tastypie==0.14.7
djangorestframework==3.14.0
Flask==3.0.3
flask-paginate==2024.4.12
gunicorn==23.0.0
h11==0.14.0
httpie==3.2.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mysql-connector==2.2.9
mysqlclient==2.2.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
PyMySQL==1.1.1
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-mimeparse==2.0.0
pytz==2024.2
redis==5.0.8
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
selenium==4.24.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tomlkit==0.13.2
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
zope.interface==7.0.3
